# docs/openapi/swagger.yaml
openapi: 3.1.0
info:
  title: FGD (Federation Governance Dashboard) API
  version: "2.1.0"
  description: |
    REST interface for FGD, the Minecraft NPC swarm management system.
    Routes are mounted as follows:
      - Unversioned surface: `/api/*` — currently served by the v1 route
        tree. `/api/mineflayer` is a special case that always resolves to
        the policy-gated v2 Mineflayer surface (or 503 if the policy
        service failed to initialize).
      - `/api/v1/*` — legacy direct-control surface. Mineflayer endpoints
        under `/api/v1/mineflayer` are DEPRECATED (sunset 2026-07-01).
      - `/api/v2/*` — current surface, including the policy-gated
        Mineflayer routes at `/api/v2/mineflayer`.

    Authentication: protected routes accept either a Bearer JWT
    (`Authorization: Bearer <token>`) or an API key (`X-API-Key`).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/command/{botId}/mine:
    parameters:
      - in: path
        name: botId
        required: true
        schema: { type: string }
    post:
      tags: [Commands]
      summary: Mine a target block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                block: { type: string, maxLength: 64 }
                count: { type: integer, minimum: 1, maximum: 1000 }
              required: [block]
      responses:
        '200': { description: Mined }
        '400': { description: Validation error }
  /api/command/{botId}/craft:
    parameters:
      - in: path
        name: botId
        required: true
        schema: { type: string }
    post:
      tags: [Commands]
      summary: Craft a recipe
      responses:
        '200': { description: Crafted }
  /api/command/{botId}/attack:
    parameters:
      - in: path
        name: botId
        required: true
        schema: { type: string }
    post:
      tags: [Commands]
      summary: Attack a target entity
      responses:
        '200': { description: Attacked }
  /api/command/{botId}/explore:
    parameters:
      - in: path
        name: botId
        required: true
        schema: { type: string }
    post:
      tags: [Commands]
      summary: Explore an area with a radius
      responses:
        '200': { description: Exploring }
  /api/command/{botId}/build:
    parameters:
      - in: path
        name: botId
        required: true
        schema: { type: string }
    post:
      tags: [Commands]
      summary: Execute a blueprint of placements
      responses:
        '200': { description: Building }

  # ----- Control (interactive sessions) --------------------------------
  /api/control/{botId}/session/start:
    parameters:
      - in: path
        name: botId
        required: true
        schema: { type: string }
    post:
      tags: [Control]
      summary: Open a realtime control session for a bot
      responses:
        '200': { description: Session started }
  /api/control/{botId}/session/stop:
    parameters:
      - in: path
        name: botId
        required: true
        schema: { type: string }
    post:
      tags: [Control]
      summary: Close the realtime control session
      responses:
        '200': { description: Session stopped }
  /api/control/{botId}/state:
    parameters:
      - in: path
        name: botId
        required: true
        schema: { type: string }
    get:
      tags: [Control]
      summary: Current control-session state
      responses:
        '200': { description: State }

  # ----- LLM -----------------------------------------------------------
...